requirements.txt
asgiref==3.6.0
astroid==2.15.2
attrs==22.2.0
autoflake==2.0.2
bandit==1.7.5
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
Deprecated==1.2.14
dill==0.3.6
Django==4.2.1
django-oauth-toolkit==2.3.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
exceptiongroup==1.1.1
flake8==6.0.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.0
lazy-object-proxy==1.9.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
mypy==1.2.0
mypy-extensions==1.0.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
pathspec==0.11.1
pbr==5.11.1
platformdirs==3.2.0
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.14.0
pylint==2.17.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytest==7.2.2
pytz==2023.3
pyupgrade==3.3.1
PyYAML==6.0
requests==2.30.0
rich==13.3.3
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
serializers==0.2.4
simplejson==3.19.1
smmap==5.0.0
sqlparse==0.4.3
stevedore==5.0.0
tokenize-rt==5.0.0
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.0.2
wrapt==1.15.0
